feat: creation of base clases and src module
