+++ src/baseHandlers/Controller.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class Base_Controller(ABC):
+    async def handle(self):
+        pass
