+++ src/baseHandlers/Use_Case.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class Base_Use_Case(ABC):
+    async def execute(self):
+        pass
