+++ src/users_api/dtos/update_user_dto.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel, EmailStr
+
+
+
+class Update_User_DTO(BaseModel):
+    name: str|None  = None
+    email: EmailStr | None = None
+    password: str | None = None
